+++ controllers/editController.js
@@ -1,4 +1,5 @@
 const asyncHandler = require("express-async-handler");
+const { validationResult } = require("express-validator");
 const {
   getItem,
   getItems,
@@ -7,6 +8,9 @@ const {
   updateItems,
   getCategory,
 } = require("../db/queries");
+const validateCategory = require("../validators/categoryValidator");
+const validatePassword = require("../validators/passwordValidator");
+const validateItem = require("../validators/itemValidator");
 
 const editController = {
   getEditCategory: asyncHandler(async (req, res) => {
@@ -15,13 +19,13 @@ const editController = {
     console.log("req.params:", req.params);
     const categories = await getCategories();
     const { category } = req.params;
-    const input = { category };
-    const editCategory = true;
+    // const editCategory = true;
     res.render("editCategory", {
       title: "Edit Category",
       categories,
-      editCategory,
-      input,
+      inputs: { category },
+      password: true,
+      action: "edit",
     });
   }),
   getEditItem: asyncHandler(async (req, res) => {
@@ -31,25 +35,92 @@ const editController = {
     const categories = await getCategories();
     const item = await getItem(req.params);
     console.log(item);
-    const editItem = true;
     res.render("editItem", {
       title: "Edit Item",
       categories,
-      editItem,
       inputs: { ...item },
+      password: true,
+      action: "edit",
     });
   }),
-  postEditCategory: asyncHandler(async (req, res) => {
-    console.log("postEditCategory running...");
-    console.log("req.url:", req.url);
-    console.log("req.query:", req.query);
-    console.log("req.params:", req.params);
-    console.log("req.body:", req.body);
-    // If category does not change
-    // If category is empty
-    // If category is valid
-  }),
-  postEditItem: asyncHandler(async (req, res) => {}),
+  postEditCategory: [
+    validateCategory,
+    validatePassword,
+    asyncHandler(async (req, res) => {
+      console.log("postEditCategory running...");
+      const errors = validationResult(req);
+
+      if (!errors.isEmpty()) {
+        const categories = await getCategories();
+        console.log(errors);
+        const localErrors = errors
+          .array()
+          .reduce((accumulator, currentError) => {
+            const { path, value, msg } = currentError;
+            return { ...accumulator, [path]: { value, msg } };
+          }, {});
+
+        console.log(localErrors);
+        return res.status(400).render("editCategory", {
+          title: "Edit Category",
+          errors: { ...localErrors },
+          inputs: { ...req.body },
+          categories,
+          password: true,
+          action: "edit",
+        });
+      }
+
+      // await insertCategory(req.body);
+      const categories = await getCategories();
+      res.render("editCategory", {
+        title: "Edit Category",
+        categories,
+        inputs: { ...req.body },
+        password: true,
+        action: "edit",
+      });
+    }),
+  ],
+  postEditItem: [
+    validateItem,
+    validatePassword,
+    asyncHandler(async (req, res) => {
+      console.log("postEditCategory running...");
+      const errors = validationResult(req);
+
+      if (!errors.isEmpty()) {
+        const categories = await getCategories();
+        console.log(errors);
+        const localErrors = errors
+          .array()
+          .reduce((accumulator, currentError) => {
+            const { path, value, msg } = currentError;
+            return { ...accumulator, [path]: { value, msg } };
+          }, {});
+
+        console.log(localErrors);
+        return res.status(400).render("editItem", {
+          title: "Edit Item",
+          errors: { ...localErrors },
+          inputs: { ...req.body },
+          categories,
+          password: true,
+          action: "edit",
+        });
+      }
+
+      // await insertCategory(req.body);
+      const categories = await getCategories();
+      res.render("editItem", {
+        title: "Edit Item",
+        categories,
+        inputs: { ...req.body },
+        password: true,
+        action: "edit",
+      });
+    }),
+  ],
 };
 
 module.exports = editController;