+++ routes/deleteRouter.js
@@ -10,5 +10,7 @@ const deleteRouter = new Router();
 
 deleteRouter.get("/category/:category", getDeleteCategory);
 deleteRouter.get("/item/:upc", getDeleteItem);
+deleteRouter.post("/category/:category", postDeleteCategory);
+deleteRouter.post("/item/:upc", postDeleteItem);
 
 module.exports = deleteRouter;